// src/app/page/master/map-deskripsi-spesifikasi/map-deskripsi-spesifikasi.interface.ts
import {EventEmitter} from '@angular/core';

export interface MapDeskripsiSpesifikasi {
	kdProduk?: string;
	kdEvent?: number;

	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
	value?: any;
}